CKEditor: Massive update
Update version, use official Angular component, add typings, add table plugin, add media plugin
