+++ API/package.json
@@ -18,11 +18,11 @@
 		"bcrypt": "^3.0.0",
 		"config": "^2.0.1",
 		"cookie-parser": "^1.4.3",
-		"express": "^4.16.3",
-		"helmet": "^3.13.0",
+		"express": "^4.16.4",
+		"helmet": "3.14.0",
 		"jsonwebtoken": "^8.3.0",
 		"method-override": "^3.0.0",
-		"mongoose": "5.3.0",
+		"mongoose": "5.3.3",
 		"morgan": "1.9.1",
 		"passport": "^0.4.0",
 		"passport-jwt": "^4.0.0",
@@ -30,11 +30,10 @@
 		"sanitize-html": "1.19.1",
 		"source-map-support": "0.5.9",
 		"ua-parser-js": "^0.7.18",
-		"validator": "10.8.0",
-		"xml": "1.0.1"
+		"validator": "10.8.0"
 	},
 	"devDependencies": {
-		"@types/bcrypt": "^2.0.0",
+		"@types/bcrypt": "^3.0.0",
 		"@types/chai": "4.1.6",
 		"@types/chai-http": "3.0.5",
 		"@types/config": "0.0.34",
@@ -44,24 +43,23 @@
 		"@types/jsonwebtoken": "^7.2.8",
 		"@types/method-override": "0.0.31",
 		"@types/mocha": "^5.2.5",
-		"@types/mongoose": "5.2.17",
+		"@types/mongoose": "5.2.19",
 		"@types/morgan": "^1.7.35",
-		"@types/node": "10.11.3",
+		"@types/node": "10.11.7",
 		"@types/passport": "^0.4.6",
 		"@types/passport-jwt": "^3.0.1",
 		"@types/passport-local": "1.0.33",
 		"@types/sanitize-html": "1.18.2",
 		"@types/ua-parser-js": "^0.7.32",
 		"@types/validator": "9.4.2",
-		"@types/xml": "^1.0.2",
-		"chai": "^4.1.2",
+		"chai": "^4.2.0",
 		"chai-http": "4.2.0",
 		"cross-env": "^5.2.0",
 		"mocha": "^5.2.0",
-		"tsc-watch": "1.0.29",
+		"tsc-watch": "1.0.30",
 		"tslint": "^5.11.0",
-		"ts-loader": "5.1.1",
-		"typescript": "3.1.1",
+		"ts-loader": "5.2.2",
+		"typescript": "3.1.3",
 		"webpack": "4.20.2",
 		"webpack-cli": "3.1.2"
 	}
