+++ CKEditor/typings.d.ts
@@ -1,5 +1,4 @@
-﻿
-declare module '*.svg' {
-  const content: any;
-  export default content;
+﻿declare module '*.svg' {
+	const content: any;
+	export default content;
 }
