+++ client/package.json
@@ -19,43 +19,45 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "7.0.0-rc.0",
-		"@angular/cdk": "7.0.0-rc.0",
-		"@angular/common": "7.0.0-rc.0",
-		"@angular/compiler": "7.0.0-rc.0",
-		"@angular/core": "7.0.0-rc.0",
-		"@angular/forms": "7.0.0-rc.0",
-		"@angular/http": "7.0.0-rc.0",
-		"@angular/material": "7.0.0-rc.0",
-		"@angular/platform-browser": "7.0.0-rc.0",
-		"@angular/platform-browser-dynamic": "7.0.0-rc.0",
-		"@angular/platform-server": "7.0.0-rc.0",
-		"@angular/router": "7.0.0-rc.0",
-		"@angular/service-worker": "7.0.0-rc.0",
-		"@nguniversal/common": "6.1.0",
-		"@nguniversal/express-engine": "6.1.0",
-		"@nguniversal/module-map-ngfactory-loader": "6.1.0",
+		"@angular/animations": "7.0.0-rc.1",
+		"@angular/cdk": "7.0.0-rc.1",
+		"@angular/common": "7.0.0-rc.1",
+		"@angular/compiler": "7.0.0-rc.1",
+		"@angular/core": "7.0.0-rc.1",
+		"@angular/forms": "7.0.0-rc.1",
+		"@angular/http": "7.0.0-rc.1",
+		"@angular/material": "7.0.0-rc.1",
+		"@angular/platform-browser": "7.0.0-rc.1",
+		"@angular/platform-browser-dynamic": "7.0.0-rc.1",
+		"@angular/platform-server": "7.0.0-rc.1",
+		"@angular/router": "7.0.0-rc.1",
+		"@angular/service-worker": "7.0.0-rc.1",
+		"@ckeditor/ckeditor5-angular": "^1.0.0",
+		"@nguniversal/common": "7.0.0-rc.0",
+		"@nguniversal/express-engine": "7.0.0-rc.0",
+		"@nguniversal/module-map-ngfactory-loader": "7.0.0-rc.0",
 		"cookie-parser": "^1.4.3",
 		"core-js": "^2.5.7",
 		"date-fns": "2.0.0-alpha.20",
 		"express": "^4.16.3",
 		"hammerjs": "^2.0.8",
+		"iframely": "1.2.7",
 		"intersection-observer": "^0.5.0",
 		"rxjs": "6.3.3",
 		"zone.js": "^0.8.26"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "0.9.0-beta.4",
-		"@angular/cli": "7.0.0-rc.0",
-		"@angular/compiler-cli": "7.0.0-rc.0",
-		"@angular/language-service": "7.0.0-rc.0",
+		"@angular-devkit/build-angular": "0.9.0-rc.3",
+		"@angular/cli": "7.0.0-rc.3",
+		"@angular/compiler-cli": "7.0.0-rc.1",
+		"@angular/language-service": "7.0.0-rc.1",
 		"@types/cookie-parser": "^1.4.1",
 		"@types/express": "^4.16.0",
 		"@types/hammerjs": "2.0.36",
-		"@types/jasmine": "^2.8.8",
-		"@types/jasminewd2": "2.0.4",
-		"@types/node": "10.11.3",
-		"codelyzer": "^4.4.4",
+		"@types/jasmine": "^2.8.9",
+		"@types/jasminewd2": "2.0.5",
+		"@types/node": "10.11.7",
+		"codelyzer": "4.5.0",
 		"jasmine-core": "^3.2.1",
 		"jasmine-spec-reporter": "^4.2.1",
 		"karma": "^3.0.0",
