--- client/src/app/modules/compose-module/ckeditor-component/ckeditor.component.html
@@ -1,3 +1,0 @@
-<div class="editor">
-	<textarea #content></textarea>
-</div>
