+++ client/src/app/modules/content-module/content-controllers/dynamic.lazy.loader.ts
@@ -19,7 +19,7 @@ export abstract class DynamicLazyLoader implements OnDestroy {
 		this._ngUnsub.next();
 		this._ngUnsub.complete();
 
-		if (this.el && !this._unobserved) {
+		if (this.el && this.inter && !this._unobserved) {
 			this.inter.unobserve(this.el.nativeElement);
 		}
 	}
