+++ client/src/app/modules/content-module/content.module.ts
@@ -6,12 +6,14 @@ import { SharedModule } from '@app/modules/shared-module/shared.module';
 // Components
 import { ContentComponent } from './content-component/content.component';
 import { DynamicLinkComponent } from './content-controllers/dynamic.link.component';
+import { DynamicMediaComponent } from './content-controllers/dynamic.media.component';
 import { DynamicImageComponent } from './content-controllers/dynamic.image.component';
 
 @NgModule({
 	declarations: [
 		ContentComponent,
 		DynamicLinkComponent,
+		DynamicMediaComponent,
 		DynamicImageComponent
 	],
 	imports: [
@@ -19,6 +21,7 @@ import { DynamicImageComponent } from './content-controllers/dynamic.image.compo
 	],
 	entryComponents: [
 		DynamicLinkComponent,
+		DynamicMediaComponent,
 		DynamicImageComponent
 	],
 	exports: [
