+++ client/src/app/scss/_core.scss
@@ -48,7 +48,6 @@ a {
 	}
 }
 
-
 // Material changes
 
 .cdk-global-scrollblock {
