+++ client/src/app/services/tests/stubs.ts
@@ -2,7 +2,7 @@
 import { User, AccessRoles } from '@app/models';
 
 import {
-	AdminService, AuthService, CMSService, ContentService, HttpService, InterceptorService,
+	AdminService, AuthService, CMSService, HttpService, InterceptorService,
 	IntersectionService, LoadingService, MobileService
 } from '@app/services';
